+++ ezdxf/audit.py
@@ -6,20 +6,12 @@
 audit(drawing, stream): check a DXF drawing for errors.
 """
 from __future__ import unicode_literals
-
 from ezdxf.lldxf.types import is_pointer_code
-__author__ = "mozman <mozman@gmx.at>"
-__all__ = ['audit']
+from ezdxf.lldxf.const import Error
+from ezdxf.lldxf.validator import is_valid_layer_name
+from ezdxf.dxfentity import DXFEntity
 
 REQUIRED_ROOT_DICT_ENTRIES = ('ACAD_GROUP', 'ACAD_PLOTSTYLENAME')
-
-MISSING_REQUIRED_ROOT_DICT_ENTRY = 1
-UNDEFINED_LINETYPE = 2
-INVALID_LAYER_NAME = 3
-DUPLICATE_TABLE_ENTRY_NAME = 4
-POINTER_TARGET_NOT_EXISTS = 5
-
-INVALID_LAYER_NAME_CHARACTERS = frozenset(['<', '>', '/', '\\',  '"', ':', ';', '?', '*', '|', '=', "'"])
 
 
 class ErrorEntry(object):
@@ -36,10 +28,15 @@ def get_pointers(tags):
             yield tag.value
 
 
-class Audit(object):
+class Auditor(object):
     def __init__(self, drawing):
         self.drawing = drawing
         self.errors = []
+        self.undefined_targets = set()
+
+    def reset(self):
+        self.errors = []
+        self.undefined_targets = set()
 
     def __len__(self):
         return len(self.errors)
@@ -54,16 +51,6 @@ class Audit(object):
         for error in self.errors:
             stream.write(error.message)
 
-    def run(self):
-        dxfversion = self.drawing.dxfversion
-        if dxfversion > 'AC1009':  # modern style DXF13 or later
-            self.check_root_dict()
-        self.check_pointer_target_exists()
-        self.check_table_entries()
-        self.check_linetypes_exists()
-        self.check_dim_styles_exists()
-        self.check_for_invalid_layer_names()
-
     def add_error(self, code, message='', dxf_entity=None, data=None):
         error = ErrorEntry(code, message, dxf_entity, data)
         self.errors.append(error)
@@ -71,35 +58,27 @@ class Audit(object):
     def filter_errors(self, code):
         return (error for error in self.errors if error.code == code)
 
+    def run(self):
+        self.reset()
+        dxfversion = self.drawing.dxfversion
+        if dxfversion > 'AC1009':  # modern style DXF13 or later
+            self.check_root_dict()
+        self.check_table_entries()
+        self.check_database_entities()
+        self.check_linetypes_exists()
+        self.check_dim_styles_exists()
+        self.check_for_invalid_layer_names()
+
     def check_root_dict(self):
         root_dict = self.drawing.rootdict
         for name in REQUIRED_ROOT_DICT_ENTRIES:
             if name not in root_dict:
                 self.add_error(
-                    code=MISSING_REQUIRED_ROOT_DICT_ENTRY,
+                    code=Error.MISSING_REQUIRED_ROOT_DICT_ENTRY,
                     message='Missing root dict entry: {}'.format(name),
                     dxf_entity=root_dict,
                     data=name,
                 )
-
-    def check_pointer_target_exists(self):
-        undefined_targets = set()
-        db = self.drawing.entitydb
-        for tags in db.values():
-            for target_pointer in get_pointers(tags):
-                if target_pointer not in db:
-                    if target_pointer == '0':  # default unset pointer
-                        continue
-                    if target_pointer in undefined_targets:  # every undefined point just one time
-                        continue
-                    entity = self.drawing.get_dxf_entity(tags.get_handle())
-                    self.add_error(
-                        code=POINTER_TARGET_NOT_EXISTS,
-                        message='Pointer target does not exist. DXF entity: {}'.format(entity.dxftype()),
-                        dxf_entity=entity,
-                        data=target_pointer,
-                    )
-                    undefined_targets.add(target_pointer)
 
     def check_table_entries(self):
         self.drawing.layers.audit(self)
@@ -111,6 +90,11 @@ class Audit(object):
         self.drawing.appids.audit(self)
         self.drawing.views.audit(self)
 
+    def check_database_entities(self):
+        for handle in self.drawing.entitydb.keys():
+            entity = self.drawing.get_dxf_entity(handle)
+            entity.audit(self)
+
     def check_linetypes_exists(self):
         """
         Check for usage of undefined line types. AutoCAD does not load DXF files with undefined line types.
@@ -120,7 +104,7 @@ class Audit(object):
         def check_linetype_exists(name):
             if name not in linetypes:
                 self.add_error(
-                    code=UNDEFINED_LINETYPE,
+                    code=Error.UNDEFINED_LINETYPE,
                     message='Undefined linetype {}'.format(linetype),
                     dxf_entity=self.drawing.get_dxf_entity(handle),
                     data=linetype,
@@ -140,11 +124,13 @@ class Audit(object):
     def check_for_invalid_layer_names(self):
         """
         Check layer names for invalid characters: <>/\":;?*|='
+
+        It is not enough to check the layer table entries, because layer names can be used without a layer table entry.
         """
         def check_layer_name_is_valid(name):
-            if INVALID_LAYER_NAME_CHARACTERS.intersection(set(name)):
+            if not is_valid_layer_name(name):
                 self.add_error(
-                    code=INVALID_LAYER_NAME,
+                    code=Error.INVALID_LAYER_NAME,
                     message='Invalid layer name {}'.format(name),
                     dxf_entity=self.drawing.get_dxf_entity(handle),
                     data=name,
@@ -154,3 +140,20 @@ class Audit(object):
             layer_name = tags.noclass.get_first_value(8, None)  # layer has a fixed tag value
             if layer_name is not None:
                 check_layer_name_is_valid(layer_name)
+
+    def check_pointer_target_exists(self, entity, invalid_zero=False):
+        assert isinstance(entity, DXFEntity)
+        db = self.drawing.entitydb
+        for target_pointer in get_pointers(entity.tags):
+            if target_pointer not in db:
+                if target_pointer == '0' and not invalid_zero:  # default unset pointer
+                    continue
+                if target_pointer in self.undefined_targets:  # every undefined point just one time
+                    continue
+                self.add_error(
+                    code=Error.POINTER_TARGET_NOT_EXISTS,
+                    message='Pointer target does not exist. DXF entity: {}'.format(entity.dxftype()),
+                    dxf_entity=entity,
+                    data=target_pointer,
+                )
+                self.undefined_targets.add(target_pointer)