+++ ezdxf/drawing.py
@@ -20,7 +20,7 @@ from .tools.codepage import tocodepage, toencoding
 from .sections import Sections
 from .tools.juliandate import juliandate
 from .lldxf import repair
-from .audit import Audit
+from .audit import Auditor
 
 
 class Drawing(object):
@@ -336,7 +336,7 @@ class Drawing(object):
             Auditor() object
 
         """
-        auditor = Audit(self)
+        auditor = Auditor(self)
         auditor.run()
         return auditor