+++ NEWS.rst
@@ -6,6 +6,7 @@ version 0.8.5 - ?
 
    * CHANGED: block names are case insensitive 'TEST' == 'Test' (like AutoCAD)
    * CHANGED: table entry (layer, linetype, style, dimstyle, ...) names are case insensitive 'TEST' == 'Test' (like AutoCAD)
+   * CHANGED: raises DXFInvalidLayerName() for invalid characters in layer names: ``<>/\":;?*|=\```
 
 Version 0.8.4 - 2018-01-14